+++ packages/react-components/src/components/FileUploadCards.tsx
@@ -3,13 +3,49 @@
 
 import { Icon, IconButton } from '@fluentui/react';
 import React from 'react';
-import { ActiveFileUpload } from './SendBox';
 import { _FileCard } from './FileCard';
 import { _FileCardGroup } from './FileCardGroup';
 import { extension } from './utils';
 import { iconButtonClassName } from './styles/IconButton.styles';
 import { useMemo } from 'react';
 import { useLocaleFileCardStringsTrampoline } from './utils/common';
+import { SendBoxErrorBarError } from './SendBoxErrorBar';
+
+/**
+ * Attributes required for SendBox to show file uploads like name, progress etc.
+ * @beta
+ */
+export interface ActiveFileUpload {
+  /**
+   * Unique identifier for the file upload.
+   */
+  id: string;
+
+  /**
+   * File name to be rendered for uploaded file.
+   */
+  filename: string;
+
+  /**
+   * A number between 0 and 1 indicating the progress of the upload.
+   * This is unrelated to the `uploadComplete` property.
+   * It is only used to show the progress of the upload.
+   * Progress of 1 doesn't mark the upload as complete, set the `uploadComplete`
+   * property to true to mark the upload as complete.
+   */
+  progress: number;
+
+  /**
+   * Error to be displayed to the user if the upload fails.
+   */
+  error?: SendBoxErrorBarError;
+
+  /**
+   * `true` means that the upload is completed.
+   * This is independent of the upload `progress`.
+   */
+  uploadComplete?: boolean;
+}
 
 /**
  * Strings of _FileUploadCards that can be overridden.
